+++ lib/src/data/data.dart
@@ -1,2 +1,3 @@
 export 'src/ac_cache.dart';
 export 'src/ac_calendar_repository.dart';
+export 'src/ac_default_calendar_repository.dart';