+++ lib/src/presentation/src/widgets/src/month/src/ac_month_widget.dart
@@ -37,7 +37,7 @@ class ACMonthWidget extends StatelessWidget {
         for (int i = 0; i < days.length; i++)
           LayoutId(
             id: i,
-            child: ACDayWidget(
+            child: ACCalendarDayWidget(
               dayDate: days[i],
               monthPosition: _positionFor(days[i]),
             ),