+++ lib/src/presentation/src/widgets/src/vertical/vertical.dart
@@ -1,1 +1,0 @@
-export 'src/ac_vertical_calendar.dart';